+++ src/gameobjects/rendertexture/RenderTexture.js
@@ -122,8 +122,7 @@ var RenderTexture = new Class({
         this.globalAlpha = 1;
 
         /**
-         * The HTML Canvas Element that the Render Texture is drawing to.
-         * This is only populated if Phaser is running with the Canvas Renderer.
+         * The HTML Canvas Element that the Render Texture is drawing to when using the Canvas Renderer.
          *
          * @name Phaser.GameObjects.RenderTexture#canvas
          * @type {HTMLCanvasElement}
@@ -405,7 +404,7 @@ var RenderTexture = new Class({
 
         if (this.gl)
         {
-            this.renderer.setFramebuffer(this.framebuffer);
+            this.renderer.setFramebuffer(this.framebuffer, true);
 
             var gl = this.gl;
     
@@ -413,7 +412,7 @@ var RenderTexture = new Class({
     
             gl.clear(gl.COLOR_BUFFER_BIT);
     
-            this.renderer.setFramebuffer(null);
+            this.renderer.setFramebuffer(null, true);
         }
         else
         {
@@ -438,7 +437,7 @@ var RenderTexture = new Class({
         {
             if (this.gl)
             {
-                this.renderer.setFramebuffer(this.framebuffer);
+                this.renderer.setFramebuffer(this.framebuffer, true);
 
                 var gl = this.gl;
         
@@ -446,7 +445,7 @@ var RenderTexture = new Class({
         
                 gl.clear(gl.COLOR_BUFFER_BIT);
         
-                this.renderer.setFramebuffer(null);
+                this.renderer.setFramebuffer(null, true);
             }
             else
             {
@@ -540,7 +539,7 @@ var RenderTexture = new Class({
 
         if (gl)
         {
-            this.renderer.setFramebuffer(this.framebuffer);
+            this.renderer.setFramebuffer(this.framebuffer, true);
 
             var pipeline = this.pipeline;
     
@@ -550,7 +549,7 @@ var RenderTexture = new Class({
 
             pipeline.flush();
 
-            this.renderer.setFramebuffer(null);
+            this.renderer.setFramebuffer(null, true);
 
             pipeline.projOrtho(0, pipeline.width, pipeline.height, 0, -1000.0, 1000.0);
         }
@@ -622,8 +621,8 @@ var RenderTexture = new Class({
 
             if (gl)
             {
-                this.renderer.setFramebuffer(this.framebuffer);
-    
+                this.renderer.setFramebuffer(this.framebuffer, true);
+
                 var pipeline = this.pipeline;
         
                 pipeline.projOrtho(0, this.width, 0, this.height, -1000.0, 1000.0);
@@ -632,7 +631,7 @@ var RenderTexture = new Class({
             
                 pipeline.flush();
         
-                this.renderer.setFramebuffer(null);
+                this.renderer.setFramebuffer(null, true);
         
                 pipeline.projOrtho(0, pipeline.width, pipeline.height, 0, -1000.0, 1000.0);
             }